// github/types/paths/user/following/{username}.types.ts
export type usersCheckPersonIsFollowedByAuthenticated = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: usersCheckPersonIsFollowedByAuthenticated_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
      404: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: basic_error;
          };
        };
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Note that you'll need to set `Content-Length` to zero when calling out to this endpoint. For more information, see "[HTTP verbs](https://docs.github.com/rest/guides/getting-started-with-the-rest-api#http-method)."
 *
 * OAuth app tokens and personal access tokens (classic) need the `user:follow` scope to use this endpoint.
 */
export type usersFollow = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: usersFollow_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
      404: not_found;
      422: validation_failed;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * OAuth app tokens and personal access tokens (classic) need the `user:follow` scope to use this endpoint.
 */
export type usersUnfollow = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: usersUnfollow_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      304: not_modified;
      401: requires_authentication;
      403: forbidden;
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type usersCheckPersonIsFollowedByAuthenticated_Path = {
  username: string;
};

export type usersFollow_Path = { username: string };

export type usersUnfollow_Path = { username: string };